Move the quiz submission logic to the internal namespace
