+++ includes/class-sensei-autoloader.php
@@ -264,18 +264,18 @@ class Sensei_Autoloader {
 			/**
 			 * Quiz Submission
 			 */
-			\Sensei\Quiz_Submission\Answer\Models\Answer::class => 'quiz-submission/answer/models/class-answer.php',
-			\Sensei\Quiz_Submission\Answer\Repositories\Comments_Based_Answer_Repository::class => 'quiz-submission/answer/repositories/class-comments-based-answer-repository.php',
-			\Sensei\Quiz_Submission\Answer\Repositories\Answer_Repository_Factory::class => 'quiz-submission/answer/repositories/class-answer-repository-factory.php',
-			\Sensei\Quiz_Submission\Answer\Repositories\Answer_Repository_Interface::class => 'quiz-submission/answer/repositories/class-answer-repository-interface.php',
-			\Sensei\Quiz_Submission\Grade\Models\Grade::class => 'quiz-submission/grade/models/class-grade.php',
-			\Sensei\Quiz_Submission\Grade\Repositories\Comments_Based_Grade_Repository::class => 'quiz-submission/grade/repositories/class-comments-based-grade-repository.php',
-			\Sensei\Quiz_Submission\Grade\Repositories\Grade_Repository_Factory::class => 'quiz-submission/grade/repositories/class-grade-repository-factory.php',
-			\Sensei\Quiz_Submission\Grade\Repositories\Grade_Repository_Interface::class => 'quiz-submission/grade/repositories/class-grade-repository-interface.php',
-			\Sensei\Quiz_Submission\Submission\Models\Submission::class => 'quiz-submission/submission/models/class-submission.php',
-			\Sensei\Quiz_Submission\Submission\Repositories\Comments_Based_Submission_Repository::class => 'quiz-submission/submission/repositories/class-comments-based-submission-repository.php',
-			\Sensei\Quiz_Submission\Submission\Repositories\Submission_Repository_Factory::class => 'quiz-submission/submission/repositories/class-submission-repository-factory.php',
-			\Sensei\Quiz_Submission\Submission\Repositories\Submission_Repository_Interface::class => 'quiz-submission/submission/repositories/class-submission-repository-interface.php',
+			\Sensei\Internal\Quiz_Submission\Answer\Models\Answer::class => 'internal/quiz-submission/answer/models/class-answer.php',
+			\Sensei\Internal\Quiz_Submission\Answer\Repositories\Comments_Based_Answer_Repository::class => 'internal/quiz-submission/answer/repositories/class-comments-based-answer-repository.php',
+			\Sensei\Internal\Quiz_Submission\Answer\Repositories\Answer_Repository_Factory::class => 'internal/quiz-submission/answer/repositories/class-answer-repository-factory.php',
+			\Sensei\Internal\Quiz_Submission\Answer\Repositories\Answer_Repository_Interface::class => 'internal/quiz-submission/answer/repositories/class-answer-repository-interface.php',
+			\Sensei\Internal\Quiz_Submission\Grade\Models\Grade::class => 'internal/quiz-submission/grade/models/class-grade.php',
+			\Sensei\Internal\Quiz_Submission\Grade\Repositories\Comments_Based_Grade_Repository::class => 'internal/quiz-submission/grade/repositories/class-comments-based-grade-repository.php',
+			\Sensei\Internal\Quiz_Submission\Grade\Repositories\Grade_Repository_Factory::class => 'internal/quiz-submission/grade/repositories/class-grade-repository-factory.php',
+			\Sensei\Internal\Quiz_Submission\Grade\Repositories\Grade_Repository_Interface::class => 'internal/quiz-submission/grade/repositories/class-grade-repository-interface.php',
+			\Sensei\Internal\Quiz_Submission\Submission\Models\Submission::class => 'internal/quiz-submission/submission/models/class-submission.php',
+			\Sensei\Internal\Quiz_Submission\Submission\Repositories\Comments_Based_Submission_Repository::class => 'internal/quiz-submission/submission/repositories/class-comments-based-submission-repository.php',
+			\Sensei\Internal\Quiz_Submission\Submission\Repositories\Submission_Repository_Factory::class => 'internal/quiz-submission/submission/repositories/class-submission-repository-factory.php',
+			\Sensei\Internal\Quiz_Submission\Submission\Repositories\Submission_Repository_Interface::class => 'internal/quiz-submission/submission/repositories/class-submission-repository-interface.php',
 		);
 	}
