+++ includes/internal/quiz-submission/answer/repositories/class-answer-repository-factory.php
@@ -1,0 +1,28 @@
+<?php
+/**
+ * File containing the Answer_Repository_Factory class.
+ *
+ * @package sensei
+ */
+
+namespace Sensei\Internal\Quiz_Submission\Answer\Repositories;
+
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
+
+/**
+ * Class Answer_Repository_Factory.
+ *
+ * @since $$next-version$$
+ */
+class Answer_Repository_Factory {
+	/**
+	 * Create a repository for the answers.
+	 *
+	 * @return Answer_Repository_Interface
+	 */
+	public function create(): Answer_Repository_Interface {
+		return new Comments_Based_Answer_Repository();
+	}
+}
