--- includes/quiz-submission/grade/repositories/class-grade-repository-factory.php
@@ -1,28 +1,0 @@
-<?php
-/**
- * File containing the Grade_Repository_Factory class.
- *
- * @package sensei
- */
-
-namespace Sensei\Quiz_Submission\Grade\Repositories;
-
-if ( ! defined( 'ABSPATH' ) ) {
-	exit;
-}
-
-/**
- * Class Grade_Repository_Factory.
- *
- * @since $$next-version$$
- */
-class Grade_Repository_Factory {
-	/**
-	 * Create a repository for the grades.
-	 *
-	 * @return Grade_Repository_Interface
-	 */
-	public function create(): Grade_Repository_Interface {
-		return new Comments_Based_Grade_Repository();
-	}
-}
