--- tests/unit-tests/quiz-submission/grade/repositories/test-class-grade-repository-factory.php
@@ -1,26 +1,0 @@
-<?php
-
-namespace SenseiTest\Quiz_Submission\Grade\Repositories;
-
-use Sensei\Quiz_Submission\Grade\Repositories\Grade_Repository_Interface;
-use Sensei\Quiz_Submission\Grade\Repositories\Grade_Repository_Factory;
-
-/**
- * Class Grade_Repository_Factory_Test
- *
- * @covers \Sensei\Quiz_Submission\Grade\Repositories\Grade_Repository_Factory
- */
-class Grade_Repository_Factory_Test extends \WP_UnitTestCase {
-
-	public function testCreate_WhenCalled_ReturnsGradeRepository(): void {
-		/* Arrange. */
-		$factory = new Grade_Repository_Factory();
-
-		/* Act. */
-		$actual = $factory->create();
-
-		/* Assert. */
-		self::assertInstanceOf( Grade_Repository_Interface::class, $actual );
-	}
-
-}
